backend/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-admin-interface==0.12.2
django-colorfield==0.3.0
django-cors-headers==3.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
